Conexão com o mysql usando variáveis de ambiente
